Feat: Add Viewport Framebuffer
First render to a framebuffer, then render the framebuffer to the
screen. This allows the screen resolution to be maintained efficiently.
The aspect ratio of the viewport is also now maintained irrespective of
the window size. A bit of refactoring was also done on several GFX
files.

+++ src/gfx/sprite2d.h
@@ -7,7 +7,7 @@
 
 #define SPRITES_MAX 2000
 
-bool sprite2d_init(int viewport_width, int viewport_height);
+bool sprite2d_init(unsigned long viewport_width, unsigned long viewport_height);
 void sprite2d_shutdown();
 
 void sprite2d_begin();
